--- server/db.js
@@ -1,5 +1,0 @@
-'use strict';
-
-
-
-// module.exports = mongoose.model("Room", roomSchema);
