+++ server/middlewares/index.ts
@@ -1,0 +1,5 @@
+import express from 'express';
+
+const middlewares = [express.json()]
+
+export default middlewares;
