+++ server/package.json
@@ -4,18 +4,28 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "typescript": "^5.1.6"
+    "@types/express": "^4.17.17",
+    "@typescript-eslint/eslint-plugin": "^6.2.1",
+    "@typescript-eslint/parser": "^6.2.1",
+    "eslint": "^8.46.0",
+    "eslint-plugin-react": "^7.33.1",
+    "typescript": "^5.1.6",
+    "type": "module"
   },
   "dependencies": {
+    "@types/node": "^20.4.6",
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "mongoose": "^7.4.1",
     "nodemon": "^3.0.1",
-    "socket.io": "^4.7.1"
+    "socket.io": "^4.7.2",
+    "source-map-loader": "^4.0.1",
+    "ts-loader": "^9.4.4"
   }
 }
